Attempt at applying textures
Tried to add textures to the floor and the drone cage. It appears as if
only one pixel of the texture was being applied instead of the whole
image.
Fail most likely due to the fact that those objects were editted with
Blender.
No solutions found yet. Might be impossible.